Added Constant Movement Speed
Updated the code so that the character always moves at a constant speed instead of always taking a constant time to move to different places. Before the character always took the same amount of time to move to where the player clicked, so the movement speed of the character was faster when the user clicked further away. Now the movement speed of the character is always the same.